refactor: Card utils 함수 분리, userId 변경
