+++ utils/changeDateFormat.ts
@@ -1,0 +1,5 @@
+const changeDateFormat = (time: Date): string => {
+  return Intl.DateTimeFormat("kr").format(time);
+};
+
+export default changeDateFormat;
